frontend: Rename Groups/List.react.jsx to List.tsx
